+++ pkgs/sdk/server-ai/src/Config/LdAiConfig.cs
@@ -32,32 +32,83 @@ public record LdAiConfig
         }
     }
 
+
+    /// <summary>
+    /// Information about the model provider.
+    /// </summary>
+    public record ModelProvider
+    {
+        /// <summary>
+        /// The name of the model provider.
+        /// </summary>
+        public readonly string Name;
+
+        internal ModelProvider(string name)
+        {
+            Name = name;
+        }
+    }
+
+    /// <summary>
+    /// Information about the model.
+    /// </summary>
+    public record ModelConfiguration
+    {
+        /// <summary>
+        /// The name of the model.
+        /// </summary>
+        public readonly string Name;
+
+        /// <summary>
+        /// The model's built-in parameters provided by LaunchDarkly.
+        /// </summary>
+        public readonly IReadOnlyDictionary<string, LdValue> Parameters;
+
+        /// <summary>
+        /// The model's custom parameters provided by the user.
+        /// </summary>
+        public readonly IReadOnlyDictionary<string, LdValue> Custom;
+
+        internal ModelConfiguration(string name, IReadOnlyDictionary<string, LdValue> parameters, IReadOnlyDictionary<string, LdValue> custom)
+        {
+            Name = name;
+            Parameters = parameters;
+            Custom = custom;
+        }
+    }
+
     /// <summary>
     /// Builder for constructing an LdAiConfig instance, which can be passed as the default
-    /// value to the AI Client's <see cref="LdAiClient.ModelConfig"/> method.
+    /// value to the AI Client's <see cref="LdAiClient.Config"/> method.
     /// </summary>
     public class Builder
     {
         private bool _enabled;
-        private readonly List<Message> _prompt;
-        private readonly Dictionary<string, object> _modelParams;
+        private readonly List<Message> _messages;
+        private readonly Dictionary<string, LdValue> _modelParams;
+        private readonly Dictionary<string, LdValue> _customModelParams;
+        private string _providerName;
+        private string _modelName;
 
         internal Builder()
         {
             _enabled = false;
-            _prompt = new List<Message>();
-            _modelParams = new Dictionary<string, object>();
+            _messages = new List<Message>();
+            _modelParams = new Dictionary<string, LdValue>();
+            _customModelParams = new Dictionary<string, LdValue>();
+            _providerName = "";
+            _modelName = "";
         }
 
         /// <summary>
-        /// Adds a prompt message with the given content and role. The default role is <see cref="Role.User"/>.
+        /// Adds a message with the given content and role. The default role is <see cref="Role.User"/>.
         /// </summary>
         /// <param name="content">the content, which may contain Mustache templates</param>
         /// <param name="role">the role</param>
         /// <returns>a new builder</returns>
-        public Builder AddPromptMessage(string content, Role role = Role.User)
+        public Builder AddMessage(string content, Role role = Role.User)
         {
-            _prompt.Add(new Message(content, role));
+            _messages.Add(new Message(content, role));
             return this;
         }
 
@@ -85,14 +136,48 @@ public record LdAiConfig
         }
 
         /// <summary>
-        /// Sets a parameter for the model. The value may be any object.
+        /// Sets a parameter for the model.
         /// </summary>
         /// <param name="name">the parameter name</param>
         /// <param name="value">the parameter value</param>
         /// <returns>the builder</returns>
-        public Builder SetModelParam(string name, object value)
+        public Builder SetModelParam(string name, LdValue value)
         {
             _modelParams[name] = value;
+            return this;
+        }
+
+        /// <summary>
+        /// Sets a custom parameter for the model.
+        /// </summary>
+        /// <param name="name">the custom parameter name</param>
+        /// <param name="value">the custom parameter value</param>
+        /// <returns>the builder</returns>
+        public Builder SetCustomModelParam(string name, LdValue value)
+        {
+            _customModelParams[name] = value;
+            return this;
+        }
+
+        /// <summary>
+        /// Sets the model's name. By default, this will be the empty string.
+        /// </summary>
+        /// <param name="name">the model name</param>
+        /// <returns>the builder</returns>
+        public Builder SetModelName(string name)
+        {
+            _modelName = name;
+            return this;
+        }
+
+        /// <summary>
+        /// Sets the model provider's name. By default, this will be the empty string.
+        /// </summary>
+        /// <param name="name">the name</param>
+        /// <returns>the builder</returns>
+        public Builder SetModelProviderName(string name)
+        {
+            _providerName = name;
             return this;
         }
 
@@ -102,49 +187,46 @@ public record LdAiConfig
         /// <returns>a new LdAiConfig</returns>
         public LdAiConfig Build()
         {
-            return new LdAiConfig(_enabled, _prompt, new Meta(), _modelParams);
+            return new LdAiConfig(
+                _enabled,
+                _messages,
+                new Meta(),
+                new Model
+                {
+                    Name = _modelName,
+                    Parameters = _modelParams,
+                    Custom = _customModelParams
+                },
+                new Provider{ Name = _providerName }
+            );
         }
     }
 
     /// <summary>
     /// The prompts associated with the config.
     /// </summary>
-    public readonly IReadOnlyList<Message> Prompt;
+    public readonly IReadOnlyList<Message> Messages;
 
     /// <summary>
     /// The model parameters associated with the config.
     /// </summary>
-    public readonly IReadOnlyDictionary<string, object> Model;
+    public readonly ModelConfiguration Model;
 
+    /// <summary>
+    /// Information about the model provider.
+    /// </summary>
+    public readonly ModelProvider Provider;
 
-
-    internal LdAiConfig(bool enabled, IEnumerable<Message> prompt, Meta meta, IReadOnlyDictionary<string, object> model)
+    internal LdAiConfig(bool enabled, IEnumerable<Message> messages, Meta meta, Model model, Provider provider)
     {
-        Model = model ?? new Dictionary<string, object>();
-        Prompt = prompt?.ToList() ?? new List<Message>();
-        VersionKey = meta?.VersionKey ?? "";
+        Model = new ModelConfiguration(model?.Name ?? "", model?.Parameters ?? new Dictionary<string, LdValue>(),
+            model?.Custom ?? new Dictionary<string, LdValue>());
+        Messages = messages?.ToList() ?? new List<Message>();
+        VariationKey = meta?.VariationKey ?? "";
+        Version = meta?.Version ?? 1;
         Enabled = enabled;
+        Provider = new ModelProvider(provider?.Name ?? "");
     }
-
-    private static LdValue ObjectToValue(object obj)
-    {
-        if (obj == null)
-        {
-            return LdValue.Null;
-        }
-
-        return obj switch
-        {
-            bool b => LdValue.Of(b),
-            double d => LdValue.Of(d),
-            string s => LdValue.Of(s),
-            IEnumerable<object> list => LdValue.ArrayFrom(list.Select(ObjectToValue)),
-            IDictionary<string, object> dict => LdValue.ObjectFrom(dict.ToDictionary(kv => kv.Key,
-                kv => ObjectToValue(kv.Value))),
-            _ => LdValue.Null
-        };
-    }
-
     internal LdValue ToLdValue()
     {
         return LdValue.ObjectFrom(new Dictionary<string, LdValue>
@@ -152,15 +234,25 @@ public record LdAiConfig
             { "_ldMeta", LdValue.ObjectFrom(
                 new Dictionary<string, LdValue>
                 {
-                    { "versionKey", LdValue.Of(VersionKey) },
+                    { "variationKey", LdValue.Of(VariationKey) },
+                    { "version", LdValue.Of(Version) },
                     { "enabled", LdValue.Of(Enabled) }
                 }) },
-            { "prompt", LdValue.ArrayFrom(Prompt.Select(m => LdValue.ObjectFrom(new Dictionary<string, LdValue>
+            { "messages", LdValue.ArrayFrom(Messages.Select(m => LdValue.ObjectFrom(new Dictionary<string, LdValue>
             {
                 { "content", LdValue.Of(m.Content) },
                 { "role", LdValue.Of(m.Role.ToString()) }
             }))) },
-            { "model", ObjectToValue(Model) }
+            { "model", LdValue.ObjectFrom(new Dictionary<string, LdValue>
+            {
+                { "name", LdValue.Of(Model.Name) },
+                { "parameters", LdValue.ObjectFrom(Model.Parameters) },
+                { "custom", LdValue.ObjectFrom(Model.Custom) }
+            }) },
+            {"provider", LdValue.ObjectFrom(new Dictionary<string, LdValue>
+            {
+                {"name", LdValue.Of(Provider.Name)}
+            })}
         });
     }
 
@@ -176,16 +268,18 @@ public record LdAiConfig
     /// <returns>true if enabled</returns>
     public bool Enabled { get; }
 
+    /// <summary>
+    /// This field meant for internal LaunchDarkly usage.
+    /// </summary>
+    public string VariationKey { get; }
 
     /// <summary>
     /// This field meant for internal LaunchDarkly usage.
     /// </summary>
-    public string VersionKey { get; }
+    public int Version { get; }
 
     /// <summary>
     /// Convenient helper that returns a disabled LdAiConfig.
     /// </summary>
-    public static LdAiConfig Disabled = New().Disable().Build();
-
-
+    public static LdAiConfig Disabled => New().Disable().Build();
 }