+++ pkgs/sdk/server-ai/src/Interfaces/ILdAiConfigTracker.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Threading.Tasks;
 using LaunchDarkly.Sdk.Server.Ai.Config;
-using LaunchDarkly.Sdk.Server.Ai.Provider;
+using LaunchDarkly.Sdk.Server.Ai.Tracking;
 
 namespace LaunchDarkly.Sdk.Server.Ai.Interfaces;
 
@@ -27,11 +27,21 @@ public interface ILdAiConfigTracker
     /// <summary>
     /// Tracks the duration of a task, and returns the result of the task.
     ///
+    /// If the provided task throws, then this method will also throw.
+    ///
+    /// In the case the provided function throws, this function will still
+    /// record the duration.
     /// </summary>
     /// <param name="task">the task</param>
     /// <typeparam name="T">type of the task's result</typeparam>
     /// <returns>the task</returns>
     public Task<T> TrackDurationOfTask<T>(Task<T> task);
+    
+    /// <summary>
+    /// Tracks the time it takes for the first token to be generated.
+    /// </summary>
+    /// <param name="timeToFirstTokenMs">the duration in milliseconds</param>
+    public void TrackTimeToFirstToken(float timeToFirstTokenMs);
 
     /// <summary>
     /// Tracks feedback (positive or negative) related to the output of the model.
@@ -46,8 +56,23 @@ public interface ILdAiConfigTracker
     public void TrackSuccess();
 
     /// <summary>
+    /// Tracks an unsuccessful generation event related to this config.
+    /// </summary>
+    public void TrackError();
+
+    /// <summary>
     /// Tracks a request to a provider. The request is a task that returns a <see cref="Response"/>, which
     /// contains information about the request such as token usage and metrics.
+    ///
+    /// This function will track the duration of the operation, the token
+    /// usage, and the success or error status.
+    ///
+    /// If the provided function throws, then this method will also throw.
+    ///
+    /// In the case the provided function throws, this function will record the
+    /// duration and an error.
+    ///
+    /// A failed operation will not have any token usage data.
     ///
     /// It is the responsibility of the caller to fill in the <see cref="Response"/> object with any details
     /// that should be tracked.