+++ pkgs/sdk/server-ai/test/InterpolationTests.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Text.Json;
 using System.Text.Json.Serialization;
@@ -20,12 +21,12 @@ public class InterpolationTests
         // The replacement is done this way because to use string.Format, we'd need to escape the curly braces.
         var configJson = """
                         {
-                            "_ldMeta": {"versionKey": "1", "enabled": true},
+                            "_ldMeta": {"variationKey": "1", "enabled": true},
                             "model": {},
-                            "prompt": [
+                            "messages": [
                                 {
                                     "content": "<do-not-use-in-any-tests-prompt-placeholder>",
-                                    "role": "System"
+                                    "role": "system"
                                 }
                             ]
                         }
@@ -38,9 +39,9 @@ public class InterpolationTests
         mockClient.Setup(x => x.GetLogger()).Returns(mockLogger.Object);
 
         var client = new LdAiClient(mockClient.Object);
-        var tracker = client.ModelConfig("foo", context, LdAiConfig.Disabled, variables);
+        var tracker = client.Config("foo", context, LdAiConfig.Disabled, variables);
 
-        return tracker.Config.Prompt[0].Content;
+        return tracker.Config.Messages[0].Content;
     }
 
     [Theory]
@@ -117,6 +118,37 @@ public class InterpolationTests
         Assert.Equal("hello world ! ", result);
     }
 
+    [Fact]
+    public void TestInterpolationMalformed()
+    {
+        var mockClient = new Mock<ILaunchDarklyClient>();
+        var mockLogger = new Mock<ILogger>();
+
+        const string configJson = """
+                                  {
+                                      "_ldMeta": {"variationKey": "1", "enabled": true},
+                                      "model": {},
+                                      "messages": [
+                                          {
+                                              "content": "This is a {{ malformed }]} prompt",
+                                              "role": "System"
+                                          }
+                                      ]
+                                  }
+                                  """;
+
+
+        mockClient.Setup(x =>
+            x.JsonVariation("foo", It.IsAny<Context>(), It.IsAny<LdValue>())).Returns(LdValue.Parse(configJson));
+
+        mockClient.Setup(x => x.GetLogger()).Returns(mockLogger.Object);
+
+        mockLogger.Setup(x => x.Error(It.IsAny<string>()));
+
+        var client = new LdAiClient(mockClient.Object);
+        var tracker = client.Config("foo", Context.New("key"), LdAiConfig.Disabled);
+        Assert.False(tracker.Config.Enabled);
+    }
 
     [Fact]
     public void TestInterpolationWithBasicContext()
@@ -137,5 +169,71 @@ public class InterpolationTests
             })).Build();
         var result = Eval("I can ingest over {{ ldctx.stats.power }} tokens per second!", context, null);
         Assert.Equal("I can ingest over 9000 tokens per second!", result);
+    }
+
+    [Fact]
+    public void TestInterpolationWithMultiKindContext()
+    {
+        var user = Context.Builder(ContextKind.Default, "123")
+            .Set("cat_ownership", LdValue.ObjectFrom(new Dictionary<string, LdValue>
+            {
+                { "count", LdValue.Of(12) }
+            })).Build();
+
+        var cat = Context.Builder(ContextKind.Of("cat"), "456")
+            .Set("health", LdValue.ObjectFrom(new Dictionary<string, LdValue>
+            {
+                { "hunger", LdValue.Of("off the charts") }
+            })).Build();
+
+        var context = Context.MultiBuilder().Add(user).Add(cat).Build();
+
+        var nestedVars = Eval("As an owner of {{ ldctx.user.cat_ownership.count }} cats, I must report that my cat's hunger level is {{ ldctx.cat.health.hunger }}!", context, null);
+        Assert.Equal("As an owner of 12 cats, I must report that my cat's hunger level is off the charts!", nestedVars);
+
+        var canonicalKeys = Eval("multi={{ ldctx.key }} user={{ ldctx.user.key }} cat={{ ldctx.cat.key }}", context, null);
+        Assert.Equal("multi=cat:456:user:123 user=123 cat=456", canonicalKeys);
+    }
+
+    [Fact]
+    public void TestInterpolationMultiKindDoesNotHaveAnonymousAttribute()
+    {
+        var user = Context.Builder(ContextKind.Default, "123")
+            .Set("cat_ownership", LdValue.ObjectFrom(new Dictionary<string, LdValue>
+            {
+                { "count", LdValue.Of(12) }
+            })).Build();
+
+        var cat = Context.Builder(ContextKind.Of("cat"), "456")
+            .Set("health", LdValue.ObjectFrom(new Dictionary<string, LdValue>
+            {
+                { "hunger", LdValue.Of("off the charts") }
+            })).Build();
+
+        var context = Context.MultiBuilder().Add(user).Add(cat).Build();
+
+        var result = Eval("anonymous=<{{ ldctx.anonymous }}>", context, null);
+        Assert.Equal("anonymous=<>", result);
+    }
+
+    [Fact]
+    public void TestInterpolationMultiKindContextHasKindMulti()
+    {
+        var user = Context.Builder(ContextKind.Default, "123")
+            .Set("cat_ownership", LdValue.ObjectFrom(new Dictionary<string, LdValue>
+            {
+                { "count", LdValue.Of(12) }
+            })).Build();
+
+        var cat = Context.Builder(ContextKind.Of("cat"), "456")
+            .Set("health", LdValue.ObjectFrom(new Dictionary<string, LdValue>
+            {
+                { "hunger", LdValue.Of("off the charts") }
+            })).Build();
+
+        var context = Context.MultiBuilder().Add(user).Add(cat).Build();
+
+        var result = Eval("kind={{ ldctx.kind }}", context, null);
+        Assert.Equal("kind=multi", result);
     }
 }