+++ pkgs/sdk/server-ai/test/LdAiClientTest.cs
@@ -16,7 +16,7 @@ public class LdAiClientTest
     {
         var client = new LdClientAdapter(new LdClient(Configuration.Builder("key").Offline(true).Build()));
         var aiClient = new LdAiClient(client);
-        var result= aiClient.ModelConfig("foo", Context.New("key"), LdAiConfig.Disabled);
+        var result= aiClient.Config("foo", Context.New("key"), LdAiConfig.Disabled);
         Assert.False(result.Config.Enabled);
     }
 
@@ -42,33 +42,33 @@ public class LdAiClientTest
 
         var client = new LdAiClient(mockClient.Object);
 
-        var defaultConfig = LdAiConfig.New().AddPromptMessage("Hello").Build();
+        var defaultConfig = LdAiConfig.New().AddMessage("Hello").Build();
 
-        var tracker = client.ModelConfig("foo", Context.New(ContextKind.Default, "key"), defaultConfig);
+        var tracker = client.Config("foo", Context.New(ContextKind.Default, "key"), defaultConfig);
 
         Assert.Equal(defaultConfig, tracker.Config);
     }
 
     private const string MetaDisabledExplicitly = """
                                                   {
-                                                    "_ldMeta": {"versionKey": "1", "enabled": false},
+                                                    "_ldMeta": {"variationKey": "1", "enabled": false},
                                                     "model": {},
-                                                    "prompt": []
+                                                    "messages": []
                                                   }
                                                   """;
 
     private const string MetaDisabledImplicitly = """
                                                   {
-                                                    "_ldMeta": {"versionKey": "1"},
+                                                    "_ldMeta": {"variationKey": "1"},
                                                     "model": {},
-                                                    "prompt": []
+                                                    "messages": []
                                                   }
                                                   """;
 
     private const string MissingMeta = """
                                        {
                                          "model": {},
-                                         "prompt": []
+                                         "messages": []
                                        }
                                        """;
 
@@ -95,10 +95,46 @@ public class LdAiClientTest
         // All the JSON inputs here are considered disabled, either due to lack of the 'enabled' property,
         // or if present, it is set to false. Therefore, if the default was returned, we'd see the assertion fail
         // (since calling LdAiConfig.New() constructs an enabled config by default.)
-        var tracker = client.ModelConfig("foo", Context.New(ContextKind.Default, "key"),
-            LdAiConfig.New().AddPromptMessage("foo").Build());
+        var tracker = client.Config("foo", Context.New(ContextKind.Default, "key"),
+            LdAiConfig.New().AddMessage("foo").Build());
 
         Assert.False(tracker.Config.Enabled);
+    }
+
+    [Fact]
+    public void CanSetAllDefaultValueFields()
+    {
+        var mockClient = new Mock<ILaunchDarklyClient>();
+
+        var mockLogger = new Mock<ILogger>();
+
+        mockClient.Setup(x =>
+            x.JsonVariation("foo", It.IsAny<Context>(), It.IsAny<LdValue>())).Returns(LdValue.Null);
+
+        mockClient.Setup(x => x.GetLogger()).Returns(mockLogger.Object);
+
+        var client = new LdAiClient(mockClient.Object);
+
+        var tracker = client.Config("foo", Context.New(ContextKind.Default, "key"),
+            LdAiConfig.New().
+                AddMessage("foo").
+                SetModelParam("foo", LdValue.Of("bar")).
+                SetModelName("awesome-model").
+                SetCustomModelParam("foo", LdValue.Of("baz")).
+                SetModelProviderName("amazing-provider").
+                SetEnabled(true).Build());
+
+        Assert.True(tracker.Config.Enabled);
+        Assert.Collection(tracker.Config.Messages,
+            message =>
+            {
+                Assert.Equal("foo", message.Content);
+                Assert.Equal(Role.User, message.Role);
+            });
+        Assert.Equal("amazing-provider", tracker.Config.Provider.Name);
+        Assert.Equal("bar", tracker.Config.Model.Parameters["foo"].AsString);
+        Assert.Equal("baz", tracker.Config.Model.Custom["foo"].AsString);
+        Assert.Equal("awesome-model", tracker.Config.Model.Name);
     }
 
     [Fact]
@@ -110,9 +146,8 @@ public class LdAiClientTest
 
         const string json = """
                             {
-                              "_ldMeta": {"versionKey": "1", "enabled": true},
-                              "model": {},
-                              "prompt": [{"content": "Hello!", "role": "system"}]
+                              "_ldMeta": {"variationKey": "1", "enabled": true},
+                              "messages": [{"content": "Hello!", "role": "system"}]
                             }
                             """;
 
@@ -125,14 +160,98 @@ public class LdAiClientTest
         var client = new LdAiClient(mockClient.Object);
 
         // We shouldn't get this default.
-        var tracker = client.ModelConfig("foo", context,
-            LdAiConfig.New().AddPromptMessage("Goodbye!").Build());
+        var tracker = client.Config("foo", context,
+            LdAiConfig.New().AddMessage("Goodbye!").Build());
 
-        Assert.Collection(tracker.Config.Prompt,
+        Assert.Collection(tracker.Config.Messages,
             message =>
             {
                 Assert.Equal("Hello!", message.Content);
                 Assert.Equal(Role.System, message.Role);
             });
+
+        Assert.Equal("", tracker.Config.Provider.Name);
+        Assert.Equal("", tracker.Config.Model.Name);
+        Assert.Empty(tracker.Config.Model.Custom);
+        Assert.Empty(tracker.Config.Model.Parameters);
+    }
+
+
+    [Fact]
+    public void ModelParametersAreParsed()
+    {
+
+        var mockClient = new Mock<ILaunchDarklyClient>();
+
+        var mockLogger = new Mock<ILogger>();
+
+        const string json = """
+                            {
+                              "_ldMeta": {"variationKey": "1", "enabled": true},
+                              "model" : {
+                                "name": "model-foo",
+                                "parameters": {
+                                  "foo": "bar",
+                                  "baz": 42
+                                },
+                                "custom": {
+                                  "foo": "baz",
+                                  "baz": 43
+                                }
+                              }
+                            }
+                            """;
+
+
+        mockClient.Setup(x =>
+            x.JsonVariation("foo", It.IsAny<Context>(), It.IsAny<LdValue>())).Returns(LdValue.Parse(json));
+
+        mockClient.Setup(x => x.GetLogger()).Returns(mockLogger.Object);
+
+        var context = Context.New(ContextKind.Default, "key");
+        var client = new LdAiClient(mockClient.Object);
+
+        // We shouldn't get this default.
+        var tracker = client.Config("foo", context,
+            LdAiConfig.New().AddMessage("Goodbye!").Build());
+
+        Assert.Equal("model-foo", tracker.Config.Model.Name);
+        Assert.Equal("bar", tracker.Config.Model.Parameters["foo"].AsString);
+        Assert.Equal(42, tracker.Config.Model.Parameters["baz"].AsInt);
+        Assert.Equal("baz", tracker.Config.Model.Custom["foo"].AsString);
+        Assert.Equal(43, tracker.Config.Model.Custom["baz"].AsInt);
+    }
+
+    [Fact]
+    public void ProviderConfigIsParsed()
+    {
+
+        var mockClient = new Mock<ILaunchDarklyClient>();
+
+        var mockLogger = new Mock<ILogger>();
+
+        const string json = """
+                            {
+                              "_ldMeta": {"variationKey": "1", "enabled": true},
+                              "provider": {
+                                "name": "amazing-provider"
+                              }
+                            }
+                            """;
+
+
+        mockClient.Setup(x =>
+            x.JsonVariation("foo", It.IsAny<Context>(), It.IsAny<LdValue>())).Returns(LdValue.Parse(json));
+
+        mockClient.Setup(x => x.GetLogger()).Returns(mockLogger.Object);
+
+        var context = Context.New(ContextKind.Default, "key");
+        var client = new LdAiClient(mockClient.Object);
+
+        // We shouldn't get this default.
+        var tracker = client.Config("foo", context,
+            LdAiConfig.New().AddMessage("Goodbye!").Build());
+
+        Assert.Equal("amazing-provider", tracker.Config.Provider.Name);
     }
 }