+++ pkgs/sdk/server-ai/test/LdAiConfigTest.cs
@@ -32,12 +32,12 @@ public class LdAiConfigTest
     public void CanAddPromptMessages()
     {
         var config = LdAiConfig.New()
-            .AddPromptMessage("Hello")
-            .AddPromptMessage("World", Role.System)
-            .AddPromptMessage("!", Role.Assistant)
+            .AddMessage("Hello")
+            .AddMessage("World", Role.System)
+            .AddMessage("!", Role.Assistant)
             .Build();
 
-        Assert.Collection(config.Prompt,
+        Assert.Collection(config.Messages,
             message =>
             {
                 Assert.Equal("Hello", message.Content);
@@ -55,16 +55,37 @@ public class LdAiConfigTest
             });
     }
 
-
     [Fact]
     public void CanSetModelParams()
     {
         var config = LdAiConfig.New()
-            .SetModelParam("foo", "bar")
-            .SetModelParam("baz", 42)
+            .SetModelParam("foo", LdValue.Of("bar"))
+            .SetModelParam("baz", LdValue.Of(42))
+            .SetCustomModelParam("foo", LdValue.Of("baz"))
+            .SetCustomModelParam("baz", LdValue.Of(43))
             .Build();
 
-        Assert.Equal("bar", config.Model["foo"]);
-        Assert.Equal(42, config.Model["baz"]);
+        Assert.Equal(LdValue.Of("bar"), config.Model.Parameters["foo"]);
+        Assert.Equal(LdValue.Of(42), config.Model.Parameters["baz"]);
+
+        Assert.Equal(LdValue.Of("baz"), config.Model.Custom["foo"]);
+        Assert.Equal(LdValue.Of(43), config.Model.Custom["baz"]);
+    }
+
+    [Fact]
+    public void CanSetModelId()
+    {
+        var config = LdAiConfig.New().SetModelName("awesome-model").Build();
+        Assert.Equal("awesome-model", config.Model.Name);
+    }
+
+    [Fact]
+    public void CanSetModelProviderId()
+    {
+        var config = LdAiConfig.New()
+            .SetModelProviderName("amazing-provider")
+            .Build();
+
+        Assert.Equal("amazing-provider", config.Provider.Name);
     }
 }