+++ pkgs/sdk/server-ai/test/LdAiConfigTrackerTest.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using LaunchDarkly.Sdk.Server.Ai.Config;
 using LaunchDarkly.Sdk.Server.Ai.Interfaces;
-using LaunchDarkly.Sdk.Server.Ai.Provider;
+using LaunchDarkly.Sdk.Server.Ai.Tracking;
 using Moq;
 using Xunit;
 
@@ -42,7 +42,8 @@ namespace LaunchDarkly.Sdk.Server.Ai
             var config = LdAiConfig.Disabled;
             var data = LdValue.ObjectFrom(new Dictionary<string, LdValue>
             {
-                { "versionKey", LdValue.Of(config.VersionKey) },
+                { "variationKey", LdValue.Of(config.VariationKey) },
+                { "version", LdValue.Of(config.Version) },
                 { "configKey", LdValue.Of(flagKey) }
             });
             var tracker = new LdAiConfigTracker(mockClient.Object, flagKey, config, context);
@@ -51,6 +52,24 @@ namespace LaunchDarkly.Sdk.Server.Ai
             mockClient.Verify(x => x.Track("$ld:ai:duration:total", context, data, 1.0f), Times.Once);
         }
 
+        [Fact]
+        public void CanTrackTimeToFirstToken()
+        {
+            var mockClient = new Mock<ILaunchDarklyClient>();
+            var context = Context.New("key");
+            const string flagKey = "key";
+            var config = LdAiConfig.Disabled;
+            var data = LdValue.ObjectFrom(new Dictionary<string, LdValue>
+            {
+                { "variationKey", LdValue.Of(config.VariationKey) },
+                { "version", LdValue.Of(config.Version) },
+                { "configKey", LdValue.Of(flagKey) }
+            });
+            var tracker = new LdAiConfigTracker(mockClient.Object, flagKey, config, context);
+
+            tracker.TrackTimeToFirstToken(1.0f);
+            mockClient.Verify(x => x.Track("$ld:ai:tokens:ttf", context, data, 1.0f), Times.Once);
+        }
 
         [Fact]
         public void CanTrackSuccess()
@@ -61,13 +80,36 @@ namespace LaunchDarkly.Sdk.Server.Ai
             var config = LdAiConfig.Disabled;
             var data = LdValue.ObjectFrom(new Dictionary<string, LdValue>
             {
-                { "versionKey", LdValue.Of(config.VersionKey) },
+                { "variationKey", LdValue.Of(config.VariationKey) },
+                { "version", LdValue.Of(config.Version) },
                 { "configKey", LdValue.Of(flagKey) }
             });
 
             var tracker = new LdAiConfigTracker(mockClient.Object, flagKey, config, context);
             tracker.TrackSuccess();
             mockClient.Verify(x => x.Track("$ld:ai:generation", context, data, 1.0f), Times.Once);
+            mockClient.Verify(x => x.Track("$ld:ai:generation:success", context, data, 1.0f), Times.Once);
+        }
+
+
+        [Fact]
+        public void CanTrackError()
+        {
+            var mockClient = new Mock<ILaunchDarklyClient>();
+            var context = Context.New("key");
+            const string flagKey = "key";
+            var config = LdAiConfig.Disabled;
+            var data = LdValue.ObjectFrom(new Dictionary<string, LdValue>
+            {
+                { "variationKey", LdValue.Of(config.VariationKey) },
+                { "version", LdValue.Of(config.Version) },
+                { "configKey", LdValue.Of(flagKey) }
+            });
+
+            var tracker = new LdAiConfigTracker(mockClient.Object, flagKey, config, context);
+            tracker.TrackError();
+            mockClient.Verify(x => x.Track("$ld:ai:generation", context, data, 1.0f), Times.Once);
+            mockClient.Verify(x => x.Track("$ld:ai:generation:error", context, data, 1.0f), Times.Once);
         }
 
 
@@ -80,7 +122,8 @@ namespace LaunchDarkly.Sdk.Server.Ai
             var config = LdAiConfig.Disabled;
             var data = LdValue.ObjectFrom(new Dictionary<string, LdValue>
             {
-                { "versionKey", LdValue.Of(config.VersionKey) },
+                { "variationKey", LdValue.Of(config.VariationKey) },
+                { "version", LdValue.Of(config.Version) },
                 { "configKey", LdValue.Of(flagKey) }
             });
 
@@ -114,7 +157,8 @@ namespace LaunchDarkly.Sdk.Server.Ai
             var config = LdAiConfig.Disabled;
             var data = LdValue.ObjectFrom(new Dictionary<string, LdValue>
             {
-                { "versionKey", LdValue.Of(config.VersionKey) },
+                { "variationKey", LdValue.Of(config.VariationKey) },
+                { "version", LdValue.Of(config.Version) },
                 { "configKey", LdValue.Of(flagKey) }
             });
 
@@ -135,7 +179,8 @@ namespace LaunchDarkly.Sdk.Server.Ai
             var config = LdAiConfig.Disabled;
             var data = LdValue.ObjectFrom(new Dictionary<string, LdValue>
             {
-                { "versionKey", LdValue.Of(config.VersionKey) },
+                { "variationKey", LdValue.Of(config.VariationKey) },
+                { "version", LdValue.Of(config.Version) },
                 { "configKey", LdValue.Of(flagKey) }
             });
 
@@ -163,7 +208,8 @@ namespace LaunchDarkly.Sdk.Server.Ai
             var config = LdAiConfig.Disabled;
             var data = LdValue.ObjectFrom(new Dictionary<string, LdValue>
             {
-                { "versionKey", LdValue.Of(config.VersionKey) },
+                { "variationKey", LdValue.Of(config.VariationKey) },
+                { "version", LdValue.Of(config.Version) },
                 { "configKey", LdValue.Of(flagKey) }
             });
 
@@ -189,6 +235,8 @@ namespace LaunchDarkly.Sdk.Server.Ai
 
             var result = tracker.TrackRequest(Task.Run(() => givenResponse));
             Assert.Equal(givenResponse, result.Result);
+            mockClient.Verify(x => x.Track("$ld:ai:generation:success", context, data, 1.0f), Times.Once);
+            mockClient.Verify(x => x.Track("$ld:ai:generation", context, data, 1.0f), Times.Once);
             mockClient.Verify(x => x.Track("$ld:ai:tokens:total", context, data, 1.0f), Times.Once);
             mockClient.Verify(x => x.Track("$ld:ai:tokens:input", context, data, 2.0f), Times.Once);
             mockClient.Verify(x => x.Track("$ld:ai:tokens:output", context, data, 3.0f), Times.Once);
@@ -204,7 +252,8 @@ namespace LaunchDarkly.Sdk.Server.Ai
             var config = LdAiConfig.Disabled;
             var data = LdValue.ObjectFrom(new Dictionary<string, LdValue>
             {
-                { "versionKey", LdValue.Of(config.VersionKey) },
+                { "variationKey", LdValue.Of(config.VariationKey) },
+                { "version", LdValue.Of(config.Version) },
                 { "configKey", LdValue.Of(flagKey) }
             });
 
@@ -224,6 +273,31 @@ namespace LaunchDarkly.Sdk.Server.Ai
             var result = tracker.TrackRequest(Task.Run(() => givenResponse));
             Assert.Equal(givenResponse, result.Result);
             mockClient.Verify(x => x.Track("$ld:ai:tokens:total", context, data, 1.0f), Times.Once);
+
+            // if latency isn't provided via Statistics, then it is automatically measured.
+            mockClient.Verify(x => x.Track("$ld:ai:duration:total", context, data, It.IsAny<double>()), Times.Once);
+        }
+
+        [Fact]
+        public async Task CanTrackExceptionFromResponse()
+        {
+            var mockClient = new Mock<ILaunchDarklyClient>();
+            var context = Context.New("key");
+            const string flagKey = "key";
+            var config = LdAiConfig.Disabled;
+            var data = LdValue.ObjectFrom(new Dictionary<string, LdValue>
+            {
+                { "variationKey", LdValue.Of(config.VariationKey) },
+                { "version", LdValue.Of(config.Version) },
+                { "configKey", LdValue.Of(flagKey) }
+            });
+
+            var tracker = new LdAiConfigTracker(mockClient.Object, flagKey, config, context);
+
+            await Assert.ThrowsAsync<System.Exception>(() => tracker.TrackRequest(Task.FromException<Response>(new System.Exception("I am an exception"))));
+
+            mockClient.Verify(x => x.Track("$ld:ai:generation", context, data, 1.0f), Times.Once);
+            mockClient.Verify(x => x.Track("$ld:ai:generation:error", context, data, 1.0f), Times.Once);
 
             // if latency isn't provided via Statistics, then it is automatically measured.
             mockClient.Verify(x => x.Track("$ld:ai:duration:total", context, data, It.IsAny<double>()), Times.Once);